Phrases now contain links to passages
This is implemented by including html in the phrase. Some simple parsing
would make this better.

Suggestion: "The room smells like [vinegar | the end] and old lace."

Then parse the "the end" part and turn it into "<a href='/the
end'>vinegar</a>". This seems easy enough.

Also pulled the opacity calculation out of the view (d'uh)